+++ src/pages/oauthPage/OauthPageStyle.tsx
@@ -15,7 +15,7 @@ export const Container = styled.div`
   text-align: center;
 `;
 
-export const Logo = styled.p`
+export const Logo = styled.h1`
   font-size: 1.5rem;
   font-weight: 700;
   top: 17.25rem;
