+++ src/pages/registerPage/SignUpPageStyle.tsx
